requirements.txt
streamlit
requests
Pillow
matplotlib
tensorflow
